Handle Windows-style library names.
git-svn-id: https://svn.salilab.org/imp/trunk@16098 e91f8c7b-0e32-4433-a9ac-4502d39a4eab

+++ modules/rmf/dependency/RMF/imp_install.sh
@@ -1,7 +1,14 @@
 #!/bin/sh
 
 mkdir -p $2/lib
-ln -sfF $1/libRMF* $1/_RMF.so $1/RMF.py $2/lib/
+if [ -e $1/_RMF.pyd ]; then
+  # Windows-style library names
+  ln -sfF $1/RMF.* $1/_RMF.* $2/lib/
+  ln -sfF $1/RMF.lib $2/lib/libRMF.lib
+else
+  # Unix-style library names
+  ln -sfF $1/libRMF* $1/_RMF.so $1/RMF.py $2/lib/
+fi
 
 mkdir -p $2/include/RMF
 ln -sfF $3/include/RMF/* $2/include/RMF/
